Move `metamask-controllers` types to `shared` directory
