+++ app/scripts/lib/PatchStore.ts
@@ -1,7 +1,7 @@
 import { createProjectLogger, getKnownPropertyNames } from '@metamask/utils';
 import { Patch } from 'immer';
 import { v4 as uuid } from 'uuid';
-import { MemStoreControllersComposedState } from '../metamask-controller-stores';
+import { MemStoreControllersComposedState as BackgroundState } from '../../../shared/types/metamask';
 import ComposableObservableStore from './ComposableObservableStore';
 import { sanitizeUIState } from './state-utils';
 
@@ -16,8 +16,8 @@ export class PatchStore {
 
   private listener: (request: {
     controllerKey: string;
-    oldState: MemStoreControllersComposedState;
-    newState: MemStoreControllersComposedState;
+    oldState: BackgroundState;
+    newState: BackgroundState;
   }) => void;
 
   constructor(observableStore: ComposableObservableStore) {
@@ -52,8 +52,8 @@ export class PatchStore {
     newState,
   }: {
     controllerKey: string;
-    oldState: MemStoreControllersComposedState;
-    newState: MemStoreControllersComposedState;
+    oldState: BackgroundState;
+    newState: BackgroundState;
   }) {
     const sanitizedNewState = sanitizeUIState(newState);
     const patches = this._generatePatches(oldState, sanitizedNewState);
@@ -81,12 +81,12 @@ export class PatchStore {
   }
 
   private _generatePatches(
-    oldState: MemStoreControllersComposedState,
-    newState: MemStoreControllersComposedState,
+    oldState: BackgroundState,
+    newState: BackgroundState,
   ): Patch[] {
-    return getKnownPropertyNames<keyof MemStoreControllersComposedState>(
-      newState,
-    ).reduce<Patch[]>((patches, controllerName) => {
+    return getKnownPropertyNames<keyof BackgroundState>(newState).reduce<
+      Patch[]
+    >((patches, controllerName) => {
       Object.keys(oldState[controllerName]).forEach((key) => {
         const oldData = oldState[controllerName][key];
         const newData = newState[controllerName][key];
